Revert "Change type of nInterleaved to unsigned int"
This reverts commit 67efb8c5888a6d8ef7e081fb259b9df718fd02e4.

+++ include/recoil/rans_encoder.h
@@ -14,7 +14,7 @@
 namespace Recoil {
     template<UnsignedType RansStateType, UnsignedType RansBitstreamType,
             BitCountType ProbBits, RansStateType RenormLowerBound, BitCountType WriteBits,
-            unsigned int NInterleaved,
+            size_t NInterleaved,
             bool RecordIntermediateStates = false, UnsignedType RansIntermediateStateType = RansStateType>
     class RansEncoder {
         using MyRans = Rans<RansStateType, RansBitstreamType, ProbBits, RenormLowerBound, WriteBits>;
